Rewrite styles with TS and move them to a separate file
